--- src/main/java/lotto/domain/generator/Generator.java
@@ -1,9 +1,0 @@
-package lotto.domain.generator;
-
-import lotto.domain.LottoNumbers;
-
-import java.util.List;
-
-public interface Generator {
-    List<LottoNumbers> generateLottoNumbersGroup(final int numberOfGenerating);
-}
