+++ src/test/java/lotto/domain/generator/LottoRandomGeneratorTest.java
@@ -8,15 +8,15 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class LottoGeneratorTest {
+class LottoRandomGeneratorTest {
     @Test
     @DisplayName("숫자 값을 입력받아, 6개의 로또 번호들을 값만큼 중복없이 생성해 반환한다.")
     void generate() {
         //given
         final int numberOfGenerating = 5;
-        final LottoGenerator lottoGenerator = new LottoGenerator();
+        final LottoGenerator lottoRandomGenerator = new LottoRandomGenerator();
         //when
-        final List<LottoNumbers> generated = lottoGenerator.generateLottoNumbersGroup(numberOfGenerating);
+        final List<LottoNumbers> generated = lottoRandomGenerator.generateLottoNumbersGroup(numberOfGenerating);
         final int actualDistinctCount = (int) generated.stream()
                 .distinct()
                 .count();
